+++ src/Components/MessageButton/MessageButton.js
@@ -1,16 +1,39 @@
 import React from 'react'
+import PropTypes from 'prop-types'
 import styled from 'styled-components'
 
 import { RootButton } from '../Button'
 
+import colors from '../../Styles/colors'
+
 const StyledButton = styled(RootButton)`
-  background: palegoldenrod;
+  background: ${props => props.bgColor};
+  color: ${props => props.textColor};
 `
 
 const MessageButton = props => {
-  const { children, ...rest } = props
+  const { colorMode, alignment, children, ...rest } = props
 
-  return <StyledButton {...rest}>{children}</StyledButton>
+  const bgColor = colors.message[colorMode].bg
+  const textColor = colors.message[colorMode].text
+
+  return (
+    <div style={{ textAlign: alignment }}>
+      <StyledButton bgColor={bgColor} textColor={textColor} {...rest}>
+        {children}
+      </StyledButton>
+    </div>
+  )
+}
+
+MessageButton.defaultProps = {
+  colorMode: 'off',
+  alignment: 'center'
+}
+
+MessageButton.propTypes = {
+  colorMode: PropTypes.oneOf(['off', 'deut', 'prot', 'tri', 'mono']),
+  alignment: PropTypes.oneOf(['left', 'center', 'right'])
 }
 
 /** @component */