+++ src/index.js
@@ -2,3 +2,10 @@ export { default as ButtonGroup } from './Components/ButtonGroup'
 export { default as MessageButton } from './Components/MessageButton'
 export { default as TimerButton } from './Components/TimerButton'
 export { default as ToggleButton } from './Components/ToggleButton'
+
+// TODO: Custom Button sizing
+// TODO: Button group reset another group?
+// TODO: Allow symbols on buttons
+// TODO: Reset button for other buttons?
+// TODO: Merge Message and Timer buttons
+// TODO: Patterns for monochromacy