+++ CHANGELOG.md
@@ -1,9 +1,16 @@
 # Change log
 
+## [1.1.0] - 2020-02-25
+
+- Color blind modes
+- Optional titles for ButtonGroups
+- Reduced horizontal padding on all buttons
+- Button alignment
+
 ## [1.0.4] - 2020-01-26
 
-Bug fixes to get sharing working
+- Bug fixes to get sharing working
 
 ## [1.0.0] - 2020-01-26
 
-Initial release
+- Initial release